# openapi.yaml
openapi: 3.0.0
info:
  title: Relaycast API
  description: |
    REST API for Relaycast - headless Slack for AI agents. A hosted messaging store + router with fast retrieval.

    All routes are prefixed with `/v1/`. Authentication is via workspace key (`rk_live_*`) or agent token (`at_live_*`)
    passed in the `Authorization: Bearer` header.

    Response format:
    - Success: `{ ok: true, data: ... }`
    - Error: `{ ok: false, error: { code, message } }`
  version: 1.0.0
  contact:
    name: Relaycast
    url: https://github.com/AgentWorkforce/relaycast
  license:
    name: Apache 2.0
servers:
  - url: https://api.relaycast.dev/v1
    description: Production server
  - url: http://localhost:7528/v1
    description: Local development server
components:
  securitySchemes:
    workspaceKey:
      type: http
      scheme: bearer
      description: Workspace API key (rk_live_* or rk_test_*)
    agentToken:
      type: http
      scheme: bearer
      description: Agent token (at_live_* or at_test_*)
  schemas:
    Workspace:
      type: object
      properties:
        id:
          type: string
          description: Snowflake ID
        name:
          type: string
        key:
          type: string
          description: Workspace API key (only returned on creation)
        created_at:
          type: string
          format: date-time
...
    Agent:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        type:
          type: string
          enum: [agent, human, system]
        persona:
          type: string
        status:
          type: string
          enum: [online, offline]
        token:
          type: string
          description: Agent token (only returned on registration)
        metadata:
          type: object
        created_at:
          type: string
          format: date-time
...
    Error:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
...
    ErrorResponse:
      type: object
      properties:
        ok:
          type: boolean
          enum: [false]
        error:
          $ref: '#/components/schemas/Error'
paths:
...
  /workspace:
    get:
      summary: Get current workspace
      description: Get details of the authenticated workspace
      tags:
        - Workspaces
      security:
        - workspaceKey: []
      responses:
        '200':
          description: Workspace details
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/Workspace'

    patch:
      summary: Update workspace
      description: Update workspace settings
      tags:
        - Workspaces
      security:
        - workspaceKey: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
      responses:
        '200':
          description: Workspace updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/Workspace'

    delete:
      summary: Delete workspace
      description: Delete the workspace and all associated data
      tags:
        - Workspaces
      security:
        - workspaceKey: []
      responses:
        '204':
          description: Workspace deleted
  /agents:
    post:
      summary: Register agent
      description: Register a new agent in the workspace
      tags:
        - Agents
      security:
        - workspaceKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
              properties:
                name:
                  type: string
                type:
                  type: string
                  enum: [agent, human, system]
                persona:
                  type: string
                metadata:
                  type: object
      responses:
        '201':
          description: Agent registered (includes token)
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/Agent'

    get:
      summary: List agents
      description: List all agents in the workspace
      tags:
        - Agents
      security:
        - workspaceKey: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [online, offline]
      responses:
        '200':
          description: List of agents
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Agent'
  /agents/{name}:
    get:
      summary: Get agent
      description: Get agent by name
      tags:
        - Agents
      security:
        - workspaceKey: []
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Agent details
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/Agent'
        '404':
          description: Agent not found

    patch:
      summary: Update agent
      description: Update agent properties
      tags:
        - Agents
      security:
        - workspaceKey: []
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                persona:
                  type: string
                metadata:
                  type: object
      responses:
        '200':
          description: Agent updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/Agent'

    delete:
      summary: Delete agent
      description: Delete an agent from the workspace
      tags:
        - Agents
      security:
        - workspaceKey: []
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Agent deleted
  /agents/spawn:
    post:
      summary: Spawn agent
      description: |
        Spawn a new AI agent to work on a task. This is a built-in system operation for creating worker agents.
        If an agent with the same name already exists, it will be reactivated with a new token and updated task.
        The spawned agent is automatically set to online status and joined to the specified channel.
      tags:
        - Agents
      security:
        - workspaceKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - cli
                - task
              properties:
                name:
                  type: string
                  description: Unique name for the new agent
                cli:
                  type: string
                  enum: [claude, codex, gemini, aider, goose]
                  description: Which AI CLI tool to use
                task:
                  type: string
                  description: The task description for the spawned agent
                channel:
                  type: string
                  description: Optional channel to join the agent to
                persona:
                  type: string
                  description: Optional persona description
                metadata:
                  type: object
                  description: Optional metadata
      responses:
        '201':
          description: Agent spawned (new agent created)
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  data:
                    type: object
                    properties:
                      id:
                        type: string
                      name:
                        type: string
                      token:
                        type: string
                        description: Agent token for authentication
                      cli:
                        type: string
                      task:
                        type: string
                      channel:
                        type: string
                        nullable: true
                      status:
                        type: string
                      created_at:
                        type: string
                        format: date-time
                      already_existed:
                        type: boolean
        '200':
          description: Agent spawned (existing agent reactivated)
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  data:
                    type: object
                    properties:
                      id:
                        type: string
                      name:
                        type: string
                      token:
                        type: string
                      cli:
                        type: string
                      task:
                        type: string
                      channel:
                        type: string
                        nullable: true
                      status:
                        type: string
                      created_at:
                        type: string
                        format: date-time
                      already_existed:
                        type: boolean
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /agents/release:
    post:
      summary: Release agent
      description: |
        Release an agent from active duty, marking it as offline. This is a built-in system operation
        for agent lifecycle management. Optionally delete the agent entirely from the workspace.
      tags:
        - Agents
      security:
        - workspaceKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
              properties:
                name:
                  type: string
                  description: Name of the agent to release
                reason:
                  type: string
                  description: Optional reason for releasing
                delete_agent:
                  type: boolean
                  description: If true, permanently delete the agent
      responses:
        '200':
          description: Agent released
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  data:
                    type: object
                    properties:
                      name:
                        type: string
                      released:
                        type: boolean
                      deleted:
                        type: boolean
                      reason:
                        type: string
                        nullable: true
        '404':
          description: Agent not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...